#104 Separando testes de data de favoritar em arquivos separados
