+++ hortum_mobile/test/views/favorites/fav_test.dart
@@ -1,11 +1,6 @@
-import 'dart:convert';
-
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hortum_mobile/data/announ_favorite_backend.dart';
-import 'package:hortum_mobile/data/announcements/announcements_backend.dart';
-import 'package:hortum_mobile/data/productor_favorite_backend.dart';
 import 'package:hortum_mobile/globals.dart';
 import 'package:hortum_mobile/views/favorites/components/select_favorite_button.dart';
 import 'package:hortum_mobile/views/favorites/fav_page.dart';
@@ -61,52 +56,6 @@ main() {
       home: FavPage(isAnnouncement: true, dio: dioMock),
     );
   }
-
-  group('AnnounDataApi methods tests', () {
-    test('Testing method getFavAnnoun of AnnounDataApi', () async {
-      actualUser.tokenAccess = 'token';
-      AnnouncementsApi announData = AnnouncementsApi(dioMock);
-      when(dioMock.get(any, options: anyNamed('options'))).thenAnswer(
-          (_) async => Response(data: response, requestOptions: null));
-      await announData.getFavAnnoun();
-      expect(announData.announcements, response['idAnunFav']);
-    });
-  });
-
-  group('ProductorFavAPI methods tests', () {
-    test('Testing method favProductor of ProductorFavAPI', () async {
-      ProductorFavAPI prodFavData = ProductorFavAPI(dioMock);
-      String responseMatcher = "Produtores favoritos atualizados com sucesso";
-      actualUser.tokenAccess = 'token';
-      when(dioMock.patch(any,
-              data: anyNamed('data'), options: anyNamed('options')))
-          .thenAnswer((_) async => Response(
-              data: jsonEncode(responseMatcher),
-              requestOptions: null,
-              statusCode: 200));
-      int responseActual = await prodFavData.favProductor('user@email.com');
-
-      expect(responseActual, 200);
-    });
-  });
-
-  group('AnnounFavAPI methods tests', () {
-    test('Testing method favAnnoun of AnnounFavAPI', () async {
-      AnnounFavAPI announFavData = AnnounFavAPI(dioMock);
-      String responseMatcher = "Anúncio atualizado com sucesso";
-      actualUser.tokenAccess = 'token';
-      when(dioMock.patch(any,
-              data: anyNamed('data'), options: anyNamed('options')))
-          .thenAnswer((_) async => Response(
-              data: jsonEncode(responseMatcher),
-              requestOptions: null,
-              statusCode: 200));
-      int responseActual =
-          await announFavData.favAnnoun('user@email.com', 'Melancia');
-
-      expect(responseActual, 200);
-    });
-  });
 
   group('Design Tests', () {
     testWidgets('Testing the announcements on FavPage',
